fetch: mark remote HEAD as public phase
To help improve the speed Mercurial operations, we shouldn't leave the whole
repository in the draft phase. The only hueristic git has for 'public' is
remote HEAD. Therefore, if remote HEAD exists then we use it to mark the
corresponding local commit as public.

For now, we'll put this behind a config flag for easier rollback.