heptapod-ci: tests for one fixed and one moving Mercurial version
This Gitlab CI configuration launches the tests of the target hg-git revision
against:

- the current head of Mercurial `stable` branch
- a fixed Mercurial version, assumed not to change anymore

We'll add more versions in a follow-up.

## `image`

All operations after cloning of the tested `hg-git` revision happen in a
fresh container started on this Docker image.

The ones we use are available on Docker Hub,
with definitions in https://dev.heptapod.net/octobus/ci-dockerfiles

## `before_script`

The `before_script` at toplevel will be executed for each job.

In later GitLab versions, we'll also have a `default` toplevel key that will
allow to specifiy a common value for `script` (and `before_script`, deprecating
its use at toplevel).

The installations performed by this `before_script` would be better provided
by a dedicated series of base Docker images, saving resources and execution
time. If that is done, we are able to remove this `before_script` altogether.

## `except`

In a first installment, we'll add a nightly scheduler after landing of this
configuration. We don't want it to relaunch tests against Mercurial versions
that don't change

## `script`

This form has the tests launcher build and use the appropriate Mercurial
revision (information about it displayed first).