windows: scp-style "URIs" are not local
GitLab and GitHub output something like this for the SSH location:

  git@github.com:hg-git/hg-git.github.com.git

Previously, cloning and pulling from them on Windows didn't work, as
the path was reinterpreted as a local path, made absolute and then
saved into `.hgrc`. The same happened on pull, even if you corrected
the definition.

I've adjust `test-networking.t` to use this type of path instead of
the `git+ssh://` scheme, as it's more likely to break.

Fixes #335.