push: fix publishing HEAD
The default for `hggit.use-phases` is to publish the remote HEAD and
tags. We want to do this on push as well, so that pushing to e.g.
`master` _immediately_ publishes the relevant changesets. However, we
run into the fact that we don't actually know or store the remote HEAD
anywhere, so we have to do an extra round-trip to obtain it.