subrepos: add support for push; drop .gitmodules on pull
Now that we can generate and update it from `.hgsub`, we no longer
need to retain `.gitmodules`. This means that we only have one source
of truth in each of the Mercurial and Git worlds, and allows us to
retain bidirectionality if-and-only-if no comments or subpaths
sections are added to `.hgsub`.